requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic-settings
torch
transformers
accelerate
bitsandbytes
Pillow
faster-whisper
aiomqtt
httpx
